+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## python-markdown2 2.4.11 (not yet released)
 
 - [pull #524] Fix angles being escaped in style blocks (issue #523)
+- [pull #527] Fix base64 images being corrupted in safe mode (issue #526)
 
 
 ## python-markdown2 2.4.10